extracted mergeComponent
