+++ src/binding/mergedComponent.ts
@@ -1,0 +1,114 @@
+import { BaseComponent } from "../components/base";
+import { BindingManager } from "./BindingManager";
+import { UnitSpec } from "vega-lite/build/src/spec";
+import { Field } from "vega-lite/build/src/channeldef";
+import { generateSignalFromAnchor } from "../components/utils";
+
+export const MERGED_SIGNAL_NAME = 'VGX_MERGED_SIGNAL_NAME'
+// Create a merged component to manage the cycle
+export function createMergedComponent(
+    node1Id: string,
+    node2Id: string,
+    channel: string,
+    bindingManager: BindingManager
+): BaseComponent {
+    const component1 = bindingManager.getComponent(node1Id);
+    const component2 = bindingManager.getComponent(node2Id);
+
+
+
+    if (!component1 || !component2) {
+        throw new Error(`Components not found: ${node1Id}, ${node2Id}`);
+    }
+
+    // Create a merged component that will manage the cycle
+    class MergedComponent extends BaseComponent {
+        mergedComponent: boolean;
+        constructor() {
+            super({});
+            this.mergedComponent = true;
+
+            // TODO create two configurations for each of the base component types, for now we'll just do the second item
+            this.schema = { [channel]: component2.schema[channel] };
+
+            this.anchors.set(`${channel}`, this.createAnchorProxy(
+                { [`${channel}`]: this.schema[`${channel}`] },
+                `${channel}`,
+                () => ({ 'absoluteValue': `${this.id}_${channel}` })
+            ));
+
+
+        }
+
+        compileComponent(inputContext: any): Partial<UnitSpec<Field>> {
+
+
+            console.log('compilingMergedComponent', inputContext)
+            // Create internal signals for each node
+            const nodeSignal = {
+                name: `${this.id}_${channel}`,
+                value: 0,
+                on: [] as any[]
+            };
+
+           
+            // TODO refactor Helper function to generate constraint application code
+            function applyConstraints(signalName: string, containerType: string): string {
+                if (containerType === 'Scalar') {
+                    return `clamp(${signalName}, min_value, max_value)`;
+                } else if (containerType === 'Range') {
+                    return `nearest(${signalName}, [min_value, max_value])`;
+                }
+                return signalName;
+            }
+
+            console.log('inputContextMErged', inputContext)
+
+
+
+
+
+
+
+            const outputSignals = Object.keys(this.schema).map(key => generateSignalFromAnchor(inputContext[key] || [], key, this.id, this.id, this.schema[key].container)).flat()
+
+            // now merge two signals:
+            const mergedSignal = {
+                name: `${this.id}_${channel}`,
+                value: 0,
+                on: [] as any[]
+            };
+
+            console.log('outputSignals', outputSignals)
+
+
+
+            // TODO:
+            // fix the update statement on mergedSignal
+            // right now they're empty. I think I need to add something for absolute value
+            // then we also need to handle the other constraints. 
+            // We may need to change merged component to compile after other components (e.g. so we don;t have random unfilled VGX_SIGNAL remaining. )            
+
+            //TODO: make not 0...
+            const updateStatements = inputContext[MERGED_SIGNAL_NAME].flat()
+            console.log('inputContext[MERGED_SIGNAL_NAME]', updateStatements)
+            for (const signal of updateStatements) {
+                mergedSignal.on.push({
+                    events: { signal: signal.name },
+                    update: signal.update
+                });
+            }
+
+            console.log('mergedSignal', mergedSignal)
+
+
+
+
+            return {
+                params: [mergedSignal]
+            };
+        }
+    }
+
+    return new MergedComponent();
+}
